del:删除旧版的BrutePage
