+++ app/renderer/src/main/src/pages/securityTool/newBrute/NewBrute.tsx
@@ -5,7 +5,8 @@ import {
     BruteExecuteExtraFormValue,
     BruteExecuteProps,
     BruteTypeTreeListProps,
-    NewBruteProps
+    NewBruteProps,
+    StartBruteParams
 } from "./NewBruteType"
 import {useControllableValue, useCreation, useMemoizedFn} from "ahooks"
 import {apiCancelStartBrute, apiGetAvailableBruteTypes, apiStartBrute, convertStartBruteParams} from "./utils"
@@ -28,7 +29,6 @@ import {randomString} from "@/utils/randomUtil"
 import cloneDeep from "lodash/cloneDeep"
 import {PluginExecuteResult} from "@/pages/plugins/operator/pluginExecuteResult/PluginExecuteResult"
 import {YakitFormDraggerContent} from "@/components/yakitUI/YakitForm/YakitForm"
-import {StartBruteParams} from "@/pages/brute/BrutePage"
 import {BrutePageInfoProps, PageNodeItemProps, usePageInfo} from "@/store/pageInfo"
 import {shallow} from "zustand/shallow"
 import {YakitRoute} from "@/routes/newRouteConstants"
